+++ packages/shared/package.json
@@ -4,10 +4,6 @@
   "private": true,
   "main": "src/index.ts",
   "types": "src/index.ts",
-  "files": [
-    "src/configs/manifest.json",
-    "src/configs/manifest.production.json"
-  ],
   "dependencies": {
     "@mantine/core": "5.9.5",
     "@mantine/hooks": "5.9.5",
