+++ packages/shared/src/index.ts
@@ -63,3 +63,4 @@ export * from './utils/fetch';
 export * from './utils/firebase';
 export * from './utils/search';
 export * from './utils/url';
+export * from './utils/extensionFile';
