+++ packages/shared/src/utils/extensionFile.ts
@@ -1,0 +1,17 @@
+import manifest from '@bypass/shared/src/configs/manifest.json';
+
+const FILE_NAME = {
+  prefix: 'bypass-links-',
+  suffix: '.zip',
+};
+
+export const getExtVersion = () => manifest.version;
+
+export const getFileNameFromVersion = (extVersion: string) =>
+  `${FILE_NAME.prefix}${extVersion}${FILE_NAME.suffix}`;
+
+export const getVersionFromFileName = (fileName: string) =>
+  fileName.substring(
+    fileName.indexOf(FILE_NAME.prefix) + FILE_NAME.prefix.length,
+    fileName.lastIndexOf(FILE_NAME.suffix)
+  );
