+++ turbo.json
@@ -1,5 +1,6 @@
 {
   "$schema": "https://turbo.build/schema.json",
+  "globalDependencies": ["$NODE_ENV", "$HOST_NAME"],
   "pipeline": {
     "dev": {
       "cache": false
